Moved entity rules classes to `bones_api_entity_rules.dart`.
