+++ lib/src/bones_api_extension.dart
@@ -6,6 +6,7 @@ import 'bones_api_base.dart';
 import 'bones_api_entity.dart';
 import 'bones_api_entity_annotation.dart';
 import 'bones_api_entity_reference.dart';
+import 'bones_api_entity_rules.dart';
 import 'bones_api_initializable.dart';
 import 'bones_api_module.dart';
