+++ lib/src/bones_api_module.dart
@@ -15,6 +15,7 @@ import 'bones_api_authentication.dart';
 import 'bones_api_base.dart';
 import 'bones_api_config.dart';
 import 'bones_api_entity.dart';
+import 'bones_api_entity_rules.dart';
 import 'bones_api_extension.dart';
 import 'bones_api_initializable.dart';
 import 'bones_api_security.dart';
